Refactor cy.createQuestionAndDashboard to a function helper - there was no good place to put it, so I also created new helpers/api directory - and I moved all helpers using cy.request (with 1 exception, where a CSV download helper also does a bunch of assertions) in there
