+++ e2e/support/helpers/api/addOrUpdateDashboardCard.ts
@@ -1,0 +1,41 @@
+import type {
+  CardId,
+  Dashboard,
+  DashboardCard,
+  DashboardId,
+} from "metabase-types/api";
+
+const DEFAULT_CARD = {
+  id: -1,
+  row: 0,
+  col: 0,
+  size_x: 11,
+  size_y: 8,
+  visualization_settings: {},
+  parameter_mappings: [],
+};
+
+export function addOrUpdateDashboardCard({
+  card_id,
+  dashboard_id,
+  card,
+}: {
+  card_id: CardId;
+  dashboard_id: DashboardId;
+  card: Partial<DashboardCard>;
+}): Cypress.Chainable<Cypress.Response<DashboardCard>> {
+  return cy
+    .request<Dashboard>("PUT", `/api/dashboard/${dashboard_id}`, {
+      dashcards: [
+        {
+          ...DEFAULT_CARD,
+          card_id,
+          ...card,
+        },
+      ],
+    })
+    .then(response => ({
+      ...response,
+      body: response.body.dashcards[0],
+    }));
+}
