+++ e2e/support/helpers/index.js
@@ -1,6 +1,6 @@
+export * from "./api";
 export * from "./e2e-action-helpers";
 export * from "./e2e-ad-hoc-question-helpers";
-export * from "./e2e-api-key-helpers";
 export * from "./e2e-bi-basics-helpers";
 export * from "./e2e-boolean-helpers";
 export * from "./e2e-cloud-helpers";
@@ -8,7 +8,6 @@ export * from "./e2e-collection-helpers";
 export * from "./e2e-command-palette-helpers";
 export * from "./e2e-custom-column-helpers";
 export * from "./e2e-dashboard-helpers";
-export * from "./e2e-data-model-helpers";
 export * from "./e2e-database-metadata-helpers";
 export * from "./e2e-dimension-list-helpers";
 export * from "./e2e-downloads-helpers";
@@ -24,7 +23,6 @@ export * from "./e2e-models-metadata-helpers";
 export * from "./e2e-notebook-helpers";
 export * from "./e2e-permissions-helpers";
 export * from "./e2e-qa-databases-helpers";
-export * from "./e2e-question-helpers";
 export * from "./e2e-relative-date-picker-helpers";
 export * from "./e2e-request-helpers";
 export * from "./e2e-search-helpers";
